requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-ollama>=0.0.1
groq>=0.4.0
python-dotenv>=1.0.0